demo_agent/requirements.txt
# D365 Demo Copilot — Dependencies

# Core
playwright>=1.49.0
pydantic>=2.6.0
python-dotenv>=1.0.0
rich>=13.7.0

# MCP SDK
mcp>=1.0.0

# Dataverse OAuth
msal>=1.28.0

# LLM Clients (configure one)
openai>=1.50.0

# Voice Narration (TTS)
edge-tts>=7.0.0

# Demo Script PDF Export
fpdf2>=2.7.0

# Testing
pytest>=7.0
pytest-asyncio>=0.21
pytest-cov>=4.0
